refactor based web db functionality into webdb, based has generic sql connect, webdb has specific functions using a sql.DB
