+++ logger/logger.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/windingtheropes/fs.alacriware/based"
+	"github.com/windingtheropes/fs.alacriware/based/webdb"
 )
 
 func LogRequest() gin.HandlerFunc {
@@ -15,14 +15,14 @@ func LogRequest() gin.HandlerFunc {
 		// don't log nonexistant routes
 		
 		if(c.Writer.Status() != 404) {
-			req := based.Request_Log {
+			req := webdb.Request_Log {
 				IP: c.ClientIP(),
 				Access_Time: time.Now().UnixMilli(),
 				Resource_Path: c.Request.URL.Path,
 				Token: c.Query("t"),
 				Code: c.Writer.Status(),
 			}
-			_, err := based.DB.LogRequest(req)
+			_, err := webdb.WDB.LogRequest(req)
 			if err != nil {
 				fmt.Println(err)
 			}
